asignaTareaUsuario -- con nullpointerexception revisar

+++ SGTI_EJB/ejbModule/stateless/UsuarioRemote.java
@@ -12,7 +12,7 @@ public interface UsuarioRemote  {
 	public boolean agregarUsuario(Usuario u);	
 	public boolean eliminarUsuario(int ced);	
 	public List<Usuario> listarUsuarios();	
-	public Usuario encontrarUsuario(int ced);
+	public Usuario encontrarUsuario(long ced);
 	public boolean actualizarUsuario(Usuario u);
 	
 }
